Table1_comparing_effects_of_IID_methods/requirements.txt
numpy==1.19.5
torchvision==0.9.1
matplotlib==3.4.2
torch==1.8.1
Pillow==8.2.0
